requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.32.3 # To fetch template metadata from template-service
Pillow==10.3.0   # For image manipulation (drawing text, resizing)
# WeasyPrint and its dependencies for PDF generation
WeasyPrint==61.2
CairoSVG==2.7.0 # Required by WeasyPrint for SVG support
tinycss2==1.3.0 # Required by WeasyPrint
html5lib==1.1 # Required by WeasyPrint
# Celery and Redis for asynchronous task queuing/results
celery==5.4.0
redis==5.0.0
python-dotenv==1.0.1